+++ src/app/Cars/page.tsx
@@ -1,0 +1,9 @@
+const page = () => {
+  return (
+    <div>
+       <p className='text-[12px] font-bold text-center pt-1'>Cars</p>
+    </div>
+  )
+}
+
+export default page
